requirements.txt
click==6.7
Flask==1.0.2
Flask-Cors==3.0.4
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
marshmallow==2.15.2
python-dotenv==0.8.2
six==1.11.0
Werkzeug==0.14.1
wheel==0.24.0
Flask-RESTful==0.3.6
flask-mongoengine==0.9.5
python-dateutil==2.7.3
PyJWT==1.6.4
requests==2.18.4
bcrypt==3.1.3
amqp==2.2.2
celery==4.1.1
redis==2.10.6